refactor: store shuffle in Util
