+++ src/index.ts
@@ -53,7 +53,6 @@ import prop from "./prop";
 import props from "./props";
 import reduce from "./reduce";
 import reduceLazy from "./reduceLazy";
-import shuffle from "./shuffle";
 import size from "./size";
 import some from "./some";
 import sort from "./sort";
@@ -127,7 +126,6 @@ export {
   props,
   reduce,
   reduceLazy,
-  shuffle,
   size,
   unless,
   some,
