+++ test/Util/shuffle.spec.ts
@@ -1,7 +1,4 @@
-import { delay, filter, map, pipe, toArray, toAsync } from "../../src";
-import shuffle from "../../src/Lazy/shuffle";
-import concurrent, { Concurrent } from "../../src/Lazy/concurrent";
-import { generatorMock } from "../utils";
+import { pipe, shuffle, toArray, toAsync } from "../../src";
 
 describe("shuffle", function () {
   describe("sync", function () {
@@ -11,7 +8,7 @@ describe("shuffle", function () {
     ])(
       "should return the given elements %s in shuffled order with length %s",
       function (input, expectedLength) {
-        const result = toArray(shuffle(input));
+        const result = shuffle(input);
         expect(result).toHaveLength(expectedLength);
 
         // Check all elements are present
@@ -24,12 +21,19 @@ describe("shuffle", function () {
     );
 
     it("should handle empty array", function () {
-      const result = toArray(shuffle([]));
+      const result = shuffle([]);
       expect(result).toEqual([]);
     });
 
+    it("should be handle immutable", function () {
+      const obj = [1, 2, 3, 4];
+      const result = shuffle(obj);
+      expect(obj !== result).toEqual(true);
+      expect(obj).toEqual([1, 2, 3, 4]);
+    });
+
     it("should handle single element", function () {
-      const result = toArray(shuffle([42]));
+      const result = shuffle([42]);
       expect(result).toEqual([42]);
     });
 
@@ -45,7 +49,7 @@ describe("shuffle", function () {
 
       // Generate multiple shuffles
       for (let i = 0; i < 50; i++) {
-        results.push(toArray(shuffle(input)));
+        results.push(shuffle(input));
       }
 
       // Check that not all results are identical (very unlikely with proper shuffling)
@@ -64,7 +68,7 @@ describe("shuffle", function () {
     ])(
       "should return the given elements %s in shuffled order with length %s",
       async function (input, expectedLength) {
-        const result = await toArray(shuffle(toAsync(input as any)));
+        const result = await shuffle(toAsync(input as unknown[]));
         expect(result).toHaveLength(expectedLength);
 
         // Check all elements are present
@@ -77,7 +81,7 @@ describe("shuffle", function () {
     );
 
     it("should handle empty async array", async function () {
-      const result = await toArray(shuffle(toAsync([])));
+      const result = await shuffle(toAsync([]));
       expect(result).toEqual([]);
     });
 
@@ -87,35 +91,13 @@ describe("shuffle", function () {
       expect(res.sort()).toEqual([1, 2, 3, 4, 5, 6]);
     });
 
-    it("should be shuffled elements concurrently", async function () {
-      const res = await pipe(
-        toAsync([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]),
-        map((a) => delay(50, a)),
-        filter((a) => a % 2 === 0),
-        shuffle,
-        concurrent(3),
-        toArray,
-      );
-      expect(res).toHaveLength(5);
-      expect(res.sort((a, b) => a - b)).toEqual([2, 4, 6, 8, 10]);
-    }, 450);
-
-    it("should be passed concurrent object when job works concurrently", async function () {
-      const mock = generatorMock();
-      const iter = shuffle(mock);
-      const concurrent = Concurrent.of(2) as any;
-
-      await iter.next(concurrent);
-      expect((mock as any).getConcurrent()).toEqual(concurrent);
-    });
-
     it("should produce different results on multiple calls (statistical test)", async function () {
       const input = [1, 2, 3, 4, 5, 6, 7, 8];
       const results = [];
 
       // Generate multiple shuffles
       for (let i = 0; i < 20; i++) {
-        results.push(await toArray(shuffle(toAsync(input))));
+        results.push(await shuffle(toAsync(input)));
       }
 
       // Check that not all results are identical (very unlikely with proper shuffling)
